Conform `Data` to `TemplateDataRepresentable` by encoding it has a hex-encoded string
Motivation: encoding arrays as `TemplateData` seems to have accidentally-quadratic runtime because a copy of the array needs to be made at each call to `NestedData.set()` (a separate issue about that will be filed soon). This PR avoids the issue for the case where a model contains a `Data` field that gets encoded by default by the `TemplateDataEncoder`. (In one case, this change has reduced render times on a large page from ~3.75s to ~460ms in a Release build.)

In addition, encoding `Data` as a hex-encoded string arguably makes more sense than encoding it as an array of integers.